\name{ildRealization.S4C}
\alias{ildRealization.S4C}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ realization of an incomplete longitudinal dataset
as an realization from an instance of class ildConf }
\description{
class is instantiated upon invocation of generic \code{sampleFrom}
on an instance of class \code{ildConf}
}
\usage{
#ildRealization.S4C(n, K, subjectModelName, subjectModelParms, 
      # subjectNonrespFun, monotone, subjectRealizer, mat, seed)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n ... monotone}{ see class ildConf }
  \item{mat}{ an instance of class longimat }
  \item{seed}{ the seed used to generate the data of the instance }
}
\details{
}
\value{
}
\references{ ~put references to the literature/web site here ~ }
\author{ ~~who you are~~ }
\note{ ~~further notes~~ }

\examples{
}
\keyword{ missing data }
\keyword{ longitudinal data }
\keyword{models}

